fix(lint): pin ruff 0.15.16 + clear newer-version findings
PR #2 went green locally on ruff 0.6.x but red on CI -- 'pip install
ruff>=0.6' resolved to 0.15.16 on the runner, which has 5 new rule
families enabled by default (UP042, UP045, RUF034, RUF046, RUF059).

This commit:

1. Pin ruff to ==0.15.16 in dev extras so local matches CI.

2. Replace UP007 with UP045 in ignore list -- the rule was renamed
   in 0.13. Same rationale (Optional on Textual generic bases).

3. Add UP042 ignore -- (str, Enum) mixin is intentional for the
   small set of value enums here; StrEnum would change the
   __format__ behavior on some downstream usages.

4. Fix real findings (no suppress):
   - RUF034 widgets/card.py: collapsed pointless '"white" if name
     else "white"' to just '"white"' -- both branches were
     identical, a vestige of an earlier refactor.
   - RUF046 music.py / music_view.py: dropped redundant int() casts
     around round() -- round(float) already returns int in Py3.
   - RUF059 test_session_coordinator: prefixed unused 'eng' tuple-
     unpack returns with '_eng' (3 sites; left the one site that
     actually uses eng untouched).

5. Re-format the four files that newer ruff format wanted reshaped
   (cosmetic; same tokens).

+++ src/pomodoro/music.py
@@ -111,7 +111,7 @@ class MusicController:
         relative skip is computed from ``current`` (the polled position). Clamped
         at 0; the seconds are rounded to a whole number for the CLI."""
         base = (current or 0.0) if relative else 0.0
-        target = max(0, int(round(base + seconds)))
+        target = max(0, round(base + seconds))
         self._run("seek", str(target))
 
     def stop(self) -> None:
